refactor: Remove e-mail configuration from default config
