+++ config/config-path-routing.json
@@ -25,7 +25,8 @@
     "files-scs:config/presets/storage-wrapper.json",
     "files-scs:config/presets/cli-params.json",
     "files-scs:config/presets/identity/solid-identity-provider.json",
-    "files-scs:config/presets/identity/email-password-interaction-policy.json"
+    "files-scs:config/presets/identity/email-password-interaction-policy.json",
+    "files-scs:config/presets/identity/email-sender.json"
   ],
   "@graph": [
     {
