+++ config/presets/identity/email-password-interaction-policy.json
@@ -178,13 +178,7 @@
               "templateFile": "./email-password-interaction/resetPasswordEmail.ejs"
             },
             "args_emailSender": {
-              "@type": "EmailSender",
-              "args_senderName": "Solid Server",
-              "args_emailConfig_name": "solidproject.org",
-              "args_emailConfig_host": "smtp.ethereal.email",
-              "args_emailConfig_port": 587,
-              "args_emailConfig_auth_user": "janick.lang@ethereal.email",
-              "args_emailConfig_auth_pass": "NYEaCsqV7aVStRCbmC"
+              "@id": "urn:solid-server:default:EmailSender"
             }
           },
           "renderHandler": {
