+++ config/presets/identity/email-sender.json
@@ -1,0 +1,10 @@
+{
+  "@context": "https://linkedsoftwaredependencies.org/bundles/npm/@solid/community-server/^0.0.0/components/context.jsonld",
+  "@graph": [
+    {
+      "comment": "This value can be used to set up an actual e-mail client",
+      "@id": "urn:solid-server:default:EmailSender",
+      "@type": "UnsupportedAsyncHandler"
+    }
+  ]
+}
