+++ test/integration/config/server-memory.json
@@ -45,6 +45,16 @@
       ]
     },
     {
+      "comment": "Use an actual e-mail sender to mock in the integration test.",
+      "@id": "urn:solid-server:default:EmailSender",
+      "@type": "BaseEmailSender",
+      "args_senderName": "Solid Server",
+      "args_emailConfig_host": "smtp.example.email",
+      "args_emailConfig_port": 587,
+      "args_emailConfig_auth_user": "alice@example.email",
+      "args_emailConfig_auth_pass": "NYEaCsqV7aVStRCbmC"
+    },
+    {
       "@id": "urn:solid-server:default:RoutingResourceStore",
       "@type": "PassthroughStore",
       "PassthroughStore:_source": {
